feat: refactor wrappedFetch Tested
